[#329] REFACTOR: 시그니처 변경으로 인한 Import 경로 변경
